requirements.txt
# MMX Agent POC - Requirements
# ==============================

# Core Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Excel Support
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Progress Bars (optional, for CLI)
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Linting and Code Quality (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Logging
colorlog>=6.7.0

# Web API (for frontend)
flask>=2.3.0
flask-cors>=4.0.0
